infrastructure/universal/__main__.py
from dotenv import load_dotenv
load_dotenv("../.env")

import core
import ecs
import iam
import security
import load_balancer

__all__ = [
    "core",
    "iam",
    "ecs",
    "security",
    "load_balancer"
]
