pyproject.toml
[project]
name = "pipelines"
version = "0.1.0"
description = ""
authors = [
    { name="Toby Drane" }
]
readme = "README.md"
dependencies = [
    "pulumi",
    "pulumi-aws",
    "pulumi-awsx",
    "pulumi_terraform",
    "prefect",
    "s3fs",
    "boto3",
    "python-dotenv"
]

[tool.setuptools]
py-modules = []